+++ crates/utils/Cargo.toml
@@ -28,7 +28,7 @@ ec4rs = { version = "1.0.2", optional = true }
 fs4 = { version = "0.7.0", optional = true }
 
 # glob
-wax = { version = "0.6.0", optional = true, features = ["miette", "walk"] }
+wax = { version = "0.6.0", optional = true, features = ["walk"] }
 
 # json
 json_comments = { version = "0.2.2", optional = true }
@@ -45,6 +45,7 @@ default = []
 editor-config = ["dep:ec4rs"]
 fs-lock = ["dep:fs4", "dep:tokio"]
 glob = ["dep:wax"]
+glob-miette = ["glob", "miette", "wax/miette"]
 miette = ["dep:miette"]
 json = ["dep:json_comments", "dep:regex", "dep:serde", "dep:serde_json"]
 toml = ["dep:toml", "dep:serde"]
